+++ pcs/cli/reports/processor.py
@@ -1,3 +1,5 @@
+from typing import List
+
 from pcs.common.reports import (
     ReportItem,
     ReportItemSeverity,
@@ -15,12 +17,20 @@ from .messages import report_item_msg_from_dto
 class ReportProcessorToConsole(ReportProcessor):
     def __init__(self, debug: bool = False) -> None:
         super().__init__()
+        self._ignore_severities: List[ReportItemSeverity] = []
         self.debug = debug
 
     def _do_report(self, report_item: ReportItem) -> None:
         report_dto = report_item.to_dto()
         msg = report_item_msg_from_dto(report_dto.message).message
         severity = report_dto.severity.level
+
+        if severity in self._ignore_severities:
+            # DEBUG overrides ignoring severities for debug reports
+            if msg and self.debug and severity == ReportItemSeverity.DEBUG:
+                print(msg)
+            return
+
         if severity == ReportItemSeverity.ERROR:
             error(
                 "{msg}{force}".format(
@@ -32,3 +42,8 @@ class ReportProcessorToConsole(ReportProcessor):
             warn(msg)
         elif msg and (self.debug or severity != ReportItemSeverity.DEBUG):
             print(msg)
+
+    def suppress_reports_of_severity(
+        self, severity_list: List[ReportItemSeverity]
+    ) -> None:
+        self._ignore_severities = list(severity_list)